Changed Mellow from a Dynamic Library to Static Library; Cleaned Warnings

+++ Mellow/src/Mellow/Events/MouseEvent.h
@@ -8,7 +8,7 @@ namespace Mellow
 	class MELLOW_API MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(int x, int y)
+		MouseMovedEvent(float x, float y)
 			: m_MouseX(x), m_MouseY(y) {}
 
 		inline float GetX() const { return m_MouseX; }
@@ -31,7 +31,7 @@ namespace Mellow
 	class MELLOW_API MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(int xOffset, int yOffset)
+		MouseScrolledEvent(float xOffset, float yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
 		inline float GetXOffset() const { return m_XOffset; }
